Move sliderGroup and slider funcs to View+Sliders.swift
